Move CPO tasks into cpo directory
